splitting components into multiple files
